added getEffectiveStyle function
